refactor: improve code quality
- split the schema to a separate file as it is bound to grow as  documentation is added.
- switched the `Program` hook to early returns to reduce nesting.
